demos/adk_with_maps_grounding/requirements.txt
google-adk
pydantic[email]
rich
google-genai
fastapi
jinja2
python-multipart
uvicorn
python-dotenv